moving static report methods to the report class
