+++ checkov/common/runners/runner_registry.py
@@ -4,7 +4,6 @@ import os
 from abc import abstractmethod
 
 from checkov.common.bridgecrew.integration_features.integration_feature_registry import integration_feature_registry
-from checkov.common.models.enums import CheckResult
 from checkov.common.output.report import Report
 from checkov.common.util import dict_utils
 from checkov.terraform.context_parsers.registry import parser_registry
@@ -46,8 +45,8 @@ class RunnerRegistry(object):
                 RunnerRegistry.enrich_report_with_guidelines(scan_report, guidelines)
             if repo_root_for_plan_enrichment:
                 enriched_resources = RunnerRegistry.get_enriched_resources(repo_root_for_plan_enrichment)
-                enriched_report = RunnerRegistry.enrich_plan_report(scan_report, enriched_resources)
-                enriched_report_with_skipped = RunnerRegistry.handle_skipped_checks(enriched_report, enriched_resources)
+                enriched_report = Report().enrich_plan_report(scan_report, enriched_resources)
+                enriched_report_with_skipped = Report().handle_skipped_checks(enriched_report, enriched_resources)
                 self.scan_reports.append(enriched_report_with_skipped)
             else:
                 self.scan_reports.append(scan_report)
@@ -152,28 +151,3 @@ class RunnerRegistry(object):
                             "skipped_checks": skipped_checks
                         }
         return enriched_resources
-
-    @staticmethod
-    def enrich_plan_report(report, enriched_resources):
-        # This enriches reports with the appropriate filepath, line numbers, and codeblock
-        for record in report.failed_checks:
-            enriched_resource = enriched_resources.get(record.resource)
-            if enriched_resource:
-                record.file_path = enriched_resource["scanned_file"]
-                record.file_line_range = enriched_resource["entity_lines_range"]
-                record.code_block = enriched_resource["entity_code_lines"]
-        return report
-
-    @staticmethod
-    def handle_skipped_checks(report, enriched_resources):
-        records_to_skip = []
-        for record in report.failed_checks:
-            resource_skips = enriched_resources[record.resource]["skipped_checks"]
-            for skip in resource_skips:
-                if record.check_id in skip["id"]: 
-                    # Remove and re-add the record to make Checkov mark it as skipped
-                    report.failed_checks.remove(record)
-                    record.check_result["result"] = CheckResult.SKIPPED
-                    record.check_result["suppress_comment"] = skip["suppress_comment"]
-                    report.add_record(record)
-        return report
